renaming math_functions.hpp to mathlib.hpp
