+++ include/universal/number/posit/posit.hpp
@@ -84,8 +84,8 @@
 #include <universal/number/posit/fdp.hpp>
 
 ///////////////////////////////////////////////////////////////////////////////////////
-/// math functions
-#include <universal/number/posit/math_functions.hpp>
+/// elementary math functions library
+#include <universal/number/posit/mathlib.hpp>
 
 ///////////////////////////////////////////////////////////////////////////////////////
 /// numerical functions
